Reimplement as simple CLI layer on top of mcpipe
- Simplify envmcp by removing the programmatic interface (that
almost certainly nobody uses) and reusing the mcpipe CLI in order
to provide the same envmcp CLI interface as before

- The context behind this is that mcpipe was created as a slightly
more sophisticated extension of envmcp, which implements the
same env-loading interface as envmcp along with additional features.
Although I switched to using mcpipe myself, I kept envmcp around
for those who only want the environment-loading features without
the other features (and for those who prefer tools to follow more of
a Unix-like philosophy of being small and coherent). However, it
doesn't seem necessary to duplicate the implementation of the
environment-loading, and mcpipe is implemented in a slightly cleaner
fashion. By reimplementing envmcp as just a thin CLI adapter for
part of mcpipe's functionality, I only need to worry about maintaining
the same features in one place and any improvements or security patches
to mcpipe will be available to envmcp without further effort

- Since this is technically a breaking change (the programmatic
interface is being removed), bump to a new major version